+++ ChainRouter.php
@@ -12,6 +12,7 @@
 namespace Symfony\Cmf\Component\Routing;
 
 use Symfony\Component\Routing\RouterInterface;
+use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Symfony\Component\Routing\Matcher\RequestMatcherInterface;
 use Symfony\Component\Routing\RequestContext;
 use Symfony\Component\Routing\RequestContextAwareInterface;
@@ -24,9 +25,7 @@ use Symfony\Component\HttpKernel\CacheWarmer\WarmableInterface;
 use Psr\Log\LoggerInterface;
 
 /**
- * ChainRouter
- *
- * Allows access to a lot of different routers.
+ * The ChainRouter allows to combine several routers to try in a defined order.
  *
  * @author Henrik Bjornskov <henrik@bjrnskov.dk>
  * @author Magnus Nordlander <magnus@e-butik.se>
@@ -34,7 +33,7 @@ use Psr\Log\LoggerInterface;
 class ChainRouter implements ChainRouterInterface, WarmableInterface
 {
     /**
-     * @var \Symfony\Component\Routing\RequestContext
+     * @var RequestContext
      */
     private $context;
 
@@ -45,17 +44,17 @@ class ChainRouter implements ChainRouterInterface, WarmableInterface
     private $routers = array();
 
     /**
-     * @var \Symfony\Component\Routing\RouterInterface[] Array of routers, sorted by priority
+     * @var RouterInterface[] Array of routers, sorted by priority
      */
     private $sortedRouters;
 
     /**
-     * @var \Symfony\Component\Routing\RouteCollection
+     * @var RouteCollection
      */
     private $routeCollection;
 
     /**
-     * @var null|\Psr\Log\LoggerInterface
+     * @var null|LoggerInterface
      */
     protected $logger;
 
@@ -78,8 +77,13 @@ class ChainRouter implements ChainRouterInterface, WarmableInterface
     /**
      * {@inheritdoc}
      */
-    public function add(RouterInterface $router, $priority = 0)
+    public function add($router, $priority = 0)
     {
+        if (!$router instanceof RouterInterface
+            && !($router instanceof RequestMatcherInterface && $router instanceof UrlGeneratorInterface)
+        ) {
+            throw new \InvalidArgumentException(sprintf('%s is not a valid router.', get_class($router)));
+        }
         if (empty($this->routers[$priority])) {
             $this->routers[$priority] = array();
         }
@@ -159,6 +163,10 @@ class ChainRouter implements ChainRouterInterface, WarmableInterface
      *
      * @param string  $url
      * @param Request $request
+     *
+     * @return array An array of parameters
+     *
+     * @throws ResourceNotFoundException If no router matched.
      */
     private function doMatch($url, Request $request = null)
     {
@@ -208,15 +216,14 @@ class ChainRouter implements ChainRouterInterface, WarmableInterface
         $debug = array();
 
         foreach ($this->all() as $router) {
-            // if $router does not implement ChainedRouterInterface and $name is not a string, continue
-            if ($name && !$router instanceof ChainedRouterInterface) {
-                if (! is_string($name)) {
-                    continue;
-                }
+            // if $router does not announce it is capable of handling
+            // non-string routes and $name is not a string, continue
+            if ($name && !is_string($name) && !$router instanceof VersatileGeneratorInterface) {
+                continue;
             }
 
-            // If $router implements ChainedRouterInterface but doesn't support this route name, continue
-            if ($router instanceof ChainedRouterInterface && !$router->supports($name)) {
+            // If $router is versatile and doesn't support this route name, continue
+            if ($router instanceof VersatileGeneratorInterface && !$router->supports($name)) {
                 continue;
             }