+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Changelog
 =========
 
+* **2014-06-06**: Updated to PSR-4 autoloading
+
 1.2.0
 -----