+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Changelog
   there was a potential information loss by re-creating the request. If you use
   DynamicRouter directly, get access to the Request object or if you are
   stand-alone create the request with Request::createFromGlobals().
+  Deprecated ChainedRouterInterface as it adds no additional information over
+  VersatileGeneratorInterface.
 
 1.3.0-RC1
 ---------